+++ eft-where-am-i/Form1.resx
@@ -121,7 +121,7 @@
   <data name="btnWhereAmI.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAAEYAAABGCAYAAABxLuKEAAAAIGNIUk0AAHomAACAhAAA+gAAAIDoAAB1
-        MAAA6mAAADqYAAAXcJy6UTwAAAAJcEhZcwAADsIAAA7CARUoSoAAAAAHdElNRQfoBhYPLTTGMyfLAAAA
+        MAAA6mAAADqYAAAXcJy6UTwAAAAJcEhZcwAADsAAAA7AAWrWiQkAAAAHdElNRQfoBhYPLTTGMyfLAAAA
         GXRFWHRTb2Z0d2FyZQB3d3cuaW5rc2NhcGUub3Jnm+48GgAAEUdJREFUeF7tmwl0VdW5x+NrX/t86w31
         1fbZVSUjkEAYBJmnACUEtFoLiIAKoqCtVVpXrX0+9akgcwAhwUIGCBkZJIQECSEhgCRkICNDAhnJcMeQ
         AUUBIf/3/869J7k3nOTe4AQs9lq/dS8355z9fb+zzz77HPZ2uVvulrvlji8t08daGddG09RhLuaBni6m
