+++ eft-where-am-i/UserControls/Settings.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace eft_where_am_i
+{
+    public partial class Settings : UserControl
+    {
+        public Settings()
+        {
+            InitializeComponent();
+        }
+    }
+}
